requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
blurhash==1.1.4
boost==0.1
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.4
cycler==0.10.0
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1602075176424/work
decorator==4.4.2
defusedxml==0.6.0
dlib @ file:///D:/bld/dlib_1607165240716/work
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.0.3
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==3.1.0
idna==2.8
imageio==2.9.0
importlib-metadata==3.3.0
ipykernel==5.4.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.13.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
Mastodon.py==1.5.1
matplotlib==3.1.0
mistune==0.8.4
mpmath==1.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
nose==1.3.7
notebook==6.1.5
numpy 
oauthlib==3.1.0
olefile 
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow 
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.4
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.2.1
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.3.22
sympy==1.7.1
tb-nightly==2.5.0a20201230
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tf-estimator-nightly==2.5.0.dev2020123001
tf-nightly==2.5.0.dev20201230
threadpoolctl==2.1.0
tifffile==2020.9.3
tk==0.1.0
torch==1.7.0
torchvision==0.8.1
tornado==6.1
tqdm==4.55.0
traitlets==4.3.3
typing-extensions 
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.0
